requirements.txt
# requirements.txt
faker
psycopg2-binary